Update and rename Rabbit.java to Grass.java
work in progress...
